requirements.txt
requests
beautifulsoup4
pandas
openpyxl
python-dateutil
